newairflow3/requirements.txt
scikit-learn==0.24.2
psycopg2-binary==2.9.1
pandas==1.1.5
numpy==1.19.5
requests==2.26.0

